Adjust system catalog's directory structure
