+++ src/frontend/src/catalog/mod.rs
@@ -27,7 +27,6 @@ pub(crate) mod catalog_service;
 pub(crate) mod column_catalog;
 pub(crate) mod database_catalog;
 pub(crate) mod index_catalog;
-pub(crate) mod pg_catalog;
 pub(crate) mod root_catalog;
 pub(crate) mod schema_catalog;
 pub(crate) mod sink_catalog;
