+++ src/frontend/src/catalog/system_catalog/pg_catalog/pg_namespace.rs
@@ -14,13 +14,15 @@
 
 use risingwave_common::types::DataType;
 
-use crate::catalog::pg_catalog::PgCatalogColumnsDef;
+use crate::catalog::system_catalog::SystemCatalogColumnsDef;
 
-/// The catalog `pg_index` contains part of the information about indexes.
-/// Ref: [`https://www.postgresql.org/docs/current/catalog-pg-class.html`]
-pub const PG_INDEX_TABLE_NAME: &str = "pg_index";
-pub const PG_INDEX_COLUMNS: &[PgCatalogColumnsDef<'_>] = &[
-    (DataType::Int32, "indexrelid"),
-    (DataType::Int32, "indrelid"),
-    (DataType::Int16, "indnatts"),
+/// The catalog `pg_namespace` stores namespaces. A namespace is the structure underlying SQL
+/// schemas: each namespace can have a separate collection of relations, types, etc. without name
+/// conflicts. Ref: [`https://www.postgresql.org/docs/current/catalog-pg-namespace.html`]
+pub const PG_NAMESPACE_TABLE_NAME: &str = "pg_namespace";
+pub const PG_NAMESPACE_COLUMNS: &[SystemCatalogColumnsDef<'_>] = &[
+    (DataType::Int32, "oid"),
+    (DataType::Varchar, "nspname"),
+    (DataType::Int32, "nspowner"),
+    (DataType::Varchar, "nspacl"),
 ];
